+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Future:
+- Improved deserialization performance.
+
 # 0.7.3:
 - Fixed an issue with empty string deserialization, caused by the `EmptyStringToDefault` code in 0.7.2.